requirements.txt
SQLAlchemy
aiohttp
python-dotenv
asyncpg